Bug fix in methods that update the database
There’s a small “feature” to GORM that affects updating fields in the
database using the “struct” method. If the update tries to set the
field’s value to 0 the value is instead interpreted as “empty” and the
update is not made. See:
http://jinzhu.me/gorm/curd.html#update-changed-fields

Switched to the “map-interface” update syntax to avoid this issue.